+++ geotrek/altimetry/views.py
@@ -1,10 +1,19 @@
+import os
+
 from django.views.generic.edit import BaseDetailView
-from django.http import HttpResponse
+from django.http import HttpResponse, Http404
+from django.core.exceptions import PermissionDenied
 from django.contrib.auth.decorators import login_required
+from django.contrib.contenttypes.models import ContentType
 from django.utils.decorators import method_decorator
+from django.shortcuts import get_object_or_404
+from django.conf import settings
+from django.views import static
 
 from mapentity.decorators import view_cache_response_content
 from mapentity.views import JSONResponseMixin, LastModifiedMixin
+
+from .models import AltimetryMixin
 
 
 class HttpSVGResponse(HttpResponse):
@@ -69,3 +78,29 @@ class ElevationArea(LastModifiedMixin, JSONResponseMixin, BaseDetailView):
     def get_context_data(self, **kwargs):
         obj = self.get_object()
         return obj.get_elevation_area()
+
+
+def serve_elevation_chart(request, model_name, pk):
+    try:
+        model = ContentType.objects.get(model=model_name).model_class()
+    except:
+        raise Http404
+    if not issubclass(model, AltimetryMixin):
+        raise Http404
+    obj = get_object_or_404(model, pk=pk)
+    if not obj.is_public():
+        if not request.user.is_authenticated():
+            raise PermissionDenied
+        if not request.user.has_perm('%s.read_%s' % (model._meta.app_label, model_name)):
+            raise PermissionDenied
+    language = request.LANGUAGE_CODE
+    obj.prepare_elevation_chart(language, request.build_absolute_uri('/'))
+    path = obj.get_elevation_chart_path(language).replace(settings.MEDIA_ROOT, '').lstrip('/')
+
+    if settings.DEBUG:
+        response = static.serve(request, path, settings.MEDIA_ROOT)
+    else:
+        response = HttpResponse()
+        response['X-Accel-Redirect'] = os.path.join(settings.MEDIA_URL_SECURE, path)
+    response["Content-Type"] = 'image/png'
+    return response