+++ geotrek/trekking/factories.py
@@ -14,11 +14,17 @@ class TrekNetworkFactory(factory.Factory):
     network = factory.Sequence(lambda n: u"network %s" % n)
 
 
-class UsageFactory(factory.Factory):
-    FACTORY_FOR = models.Usage
+class PracticeFactory(factory.Factory):
+    FACTORY_FOR = models.Practice
 
-    usage = factory.Sequence(lambda n: u"usage %s" % n)
+    name = factory.Sequence(lambda n: u"usage %s" % n)
     pictogram = dummy_filefield_as_sequence('thumbnail %s')
+
+
+class AccessibilityFactory(factory.Factory):
+    FACTORY_FOR = models.Accessibility
+
+    name = factory.Sequence(lambda n: u"accessibility %s" % n)
 
 
 class RouteFactory(factory.Factory):
@@ -81,6 +87,7 @@ class TrekFactory(TopologyFactory):
 
     route = factory.SubFactory(RouteFactory)
     difficulty = factory.SubFactory(DifficultyLevelFactory)
+    practice = factory.SubFactory(PracticeFactory)
 
 
 class TrekWithPOIsFactory(TrekFactory):