+++ geotrek/trekking/translation.py
@@ -31,8 +31,12 @@ class TrekNetworkTO(TranslationOptions):
     fields = ('network', )
 
 
-class UsageTO(TranslationOptions):
-    fields = ('usage', )
+class PracticeTO(TranslationOptions):
+    fields = ('name', )
+
+
+class AccessibilityTO(TranslationOptions):
+    fields = ('name', )
 
 
 class RouteTO(TranslationOptions):
@@ -56,7 +60,8 @@ trek_translation_to_register = [
     (trekking_models.POI, POITO),
     (trekking_models.POIType, POITypeTO),
     (trekking_models.TrekNetwork, TrekNetworkTO),
-    (trekking_models.Usage, UsageTO),
+    (trekking_models.Practice, PracticeTO),
+    (trekking_models.Accessibility, AccessibilityTO),
     (trekking_models.Route, RouteTO),
     (trekking_models.DifficultyLevel, DifficultyLevelTO),
     (trekking_models.WebLink, WebLinkTO),